enumerate volumes and snapshots

+++ qml/model/DiskListModel.h
@@ -12,15 +12,15 @@ struct BlockDevice : public BaseItem
     int depth = 0,
     int children = 0,
     bool selectable = false,
-    uint64_t size = 0,
-    uint64_t free = 0) : 
+    double size = 0,
+    double free = 0) : 
   BaseItem(name, depth, children, selectable)
   {
     m_size = size;
     m_free = free;
   }
-  uint64_t m_size = 0;
-  uint64_t m_free = 0;
+  double m_size = 0;
+  double m_free = 0;
 };
 
 using SPBlockDevice = std::shared_ptr<BlockDevice>;
